ClaReSys-Project/apps/timetable-engine/requirements.txt
grpcio==1.54.0
grpcio-tools==1.54.0
python-dateutil